Add the ability to fetch the class sections currently assigned to a teacher
We have added the ability to use PowerQueries through a PowerSchool plugin (see SwiftyPowerSchool-Plugin) and added an example function using this feature. Along the way, we discovered that the HTTP method type parameter was not being applied, so we fixed that as well. We also worked on being able to pass parameters to the API calls. While this is currently possible, the parameters are actually being encoded as a JSON payload in the HTTP body. We will be revisiting this and possibly tweaking the naming of this in the future. To make sure everything was working correctly, we added a new some new tests.