Add function for getting all enrollments for a specific section
This function takes an array of section DCIDs and will return all of the students enrollments for those sections. Due to the parameter being an array, we needed to change the parameter type in the fetchData method to accept Any. A test for the new function was also added, as well as an update to the testing parameters.