Reduce redundancy in a couple of models
As we continued to add more functions for retrieving various sets of data, we noticed that there was a lot of redundancy in the models. This commit starts the process of cleaning up the redundancy, where possible, and also better organizes the models. The tests were also updated to provide better coverage of the models.