Add an extension to several models to ensure objects are wrapped in an array
During testing we discovered that some objects are not wrapped in an array when a single object is returned. For example, if only one student is returned, it is returned as a single dictionary object. However, if more than one student is returned, they are returned as an array of dictionary objects. This inconsistency caused the JSON decoding to fail. The extension fixes this by first looking for an array and if it doesn't find one, it tries to decode a single dictionary and then wraps it in an array.
We also removed the unnecessary public statements in several models and added a test for the Student model.